Leave one out cross validation and some minor tweaks

+++ src/dictawav/kernelcanvas/kernel.rs
@@ -2,17 +2,17 @@ extern crate rand;
 
 use self::rand::distributions::{IndependentSample, Range};
 use self::rand::ThreadRng;
-use std::f32;
+use std::f64;
 
 pub struct Kernel {
-    coordinates: Vec<f32>,
+    coordinates: Vec<f64>,
 }
 
 impl Kernel {
-    pub fn check_distance_squared(&self, other: &Kernel) -> f32 {
-        let mut distance_squared = 0f32;
+    pub fn check_distance_squared(&self, other: &Kernel) -> f64 {
+        let mut distance_squared = 0f64;
         for (index, coordinate) in self.coordinates.iter().enumerate() {
-            distance_squared += (*coordinate - other.coordinates[index]).powf(2f32)
+            distance_squared += (*coordinate - other.coordinates[index]).powf(2f64)
         }
 
         distance_squared
@@ -21,13 +21,13 @@ impl Kernel {
 
 pub struct KernelBuilder {
     dimension: usize,
-    range: Range<f32>,
+    range: Range<f64>,
     random_generator: ThreadRng,
 }
 
 impl KernelBuilder {
     pub fn new(dimension: usize) -> KernelBuilder {
-        let range = Range::new(-1f32, 1f32 + f32::MIN_POSITIVE);
+        let range = Range::new(-1f64, 1f64 + f64::MIN_POSITIVE);
         let random_generator = rand::thread_rng();
 
         KernelBuilder {
@@ -49,7 +49,7 @@ impl KernelBuilder {
         }
     }
 
-    pub fn build_from_coordinates(&self, coordinates: Vec<f32>) -> Kernel {
+    pub fn build_from_coordinates(&self, coordinates: Vec<f64>) -> Kernel {
         if self.dimension != coordinates.len() {
             panic!("KernelBuilder Error: Trying to build a kernel from coordinates with different dimension!");
         }
